refactor: move example app gen scripts out
