+++ packages/create-react-native-library/src/exampleApp/addCodegenBuildScript.ts
